app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Graintrade API"
authors = ["Kostiantyn Ivashchenko <kivaschenko@protonmail.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
PyJWT = "^2.9.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
bcrypt = "^4.2.0"
aiokafka = "^0.12.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
